requirements.txt
asn1crypto==0.24.0
astroid==2.2.5
beautifulsoup4==4.7.1
certifi==2019.3.9
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.4
cryptography==3.3.2
future==0.17.1
idna==2.8
isort==4.3.17
lazy-object-proxy==1.3.1
lxml==4.6.3
mccabe==0.6.1
pycparser==2.19
pylint==2.3.1
pyparsing==2.4.7
python-telegram-bot==11.1.0
PyYAML==5.4
requests==2.25.0
requests-toolbelt==0.9.1
six==1.12.0
soupsieve==1.9.1
typed-ast==1.4.1
urllib3==1.26.5
wrapt==1.11.1
